Modified filters to be in one header and cpp as well as changed Or Filter to contain children which it ors together.
