+++ param/vision/startGateVis.yaml
@@ -12,8 +12,9 @@ open: {height: 7, iters: 1, width: 7}
 filters:
     - {Convert: {code: "BGR2HSV"}}
     - {MedianBlur: {kernel_size: 3}}
-    - {InRange: {max: {hue: 20, sat: 255, val: 255}, min: {hue: 10, sat: 200, val: 100}}}
-    - {InRange: {max: {hue: 100, sat: 210, val: 200}, min: {hue: 0, sat: 0, val: 0}}}
-    - {Or: {}}
+    - {Or:
+         - {InRange: {max: {hue: 20, sat: 255, val: 255}, min: {hue: 10, sat: 200, val: 100}}}
+         - {InRange: {max: {hue: 100, sat: 210, val: 200}, min: {hue: 0, sat: 0, val: 0}}}
+      }
     - {Close: {width: 7, height: 7, iters: 2}}
     - {Open: {width: 7, height: 7, iters: 1}}
