+++ src/vision/filters/CMakeLists.txt
@@ -1,1 +1,1 @@
-add_library(vision_filters FilterSet.cpp CloseFilter.cpp ConvertFilter.cpp InRangeFilter.cpp MedianBlurFilter.cpp OpenFilter.cpp OrFilter.cpp)
+add_library(vision_filters Filter.cpp FilterSet.cpp)
