+++ src/vision/filters/Filter.hpp
@@ -4,6 +4,7 @@
 #include <ros/ros.h>
 #include <string>
 #include <map>
+#include <vector>
 
 // The following #if is to use the correct version of the cv_bridge
 // Kinetic by default uses OpenCV3 so we don't need the custom build
@@ -19,6 +20,7 @@
 #include <boost/algorithm/string.hpp>
 
 using namespace cv;
+using std::vector;
 using std::string;
 using std::map;
 using XmlRpc::XmlRpcValue;
@@ -44,6 +46,118 @@ class Filter
 
     protected:
         string name;
+};
+
+// ------------- Close Filter ----------- //
+
+class CloseFilter : public Filter
+{
+    public:
+        CloseFilter();
+        ~CloseFilter() {};
+
+        void setParams(XmlRpcValue &params);
+        void apply(Mat &image);
+        void apply(const Mat &src, Mat &dst);
+    
+    private:
+        int iterations;
+        Size size;
+};
+
+// ------------ Convert Filter ---------- //
+
+class ConvertFilter : public Filter
+{
+    public:
+        ConvertFilter();
+        ~ConvertFilter() {};
+
+        void setParams(XmlRpcValue &params);
+
+        void apply(Mat &image);
+        void apply(const Mat &src, Mat &dst);
+
+    private:
+        int code;
+
+        const std::map<string, int> paramList {
+                                    {"BGR2HSV", cv::COLOR_BGR2HSV},
+                                    {"BGR2RGB", cv::COLOR_BGR2RGB},
+                                    {"RGB2BGR", cv::COLOR_RGB2BGR},
+                                    {"RGB2HSV", cv::COLOR_RGB2HSV},
+                                    {"BGR2GRAY", cv::COLOR_BGR2GRAY},
+                                    {"RGB2GRAY", cv::COLOR_RGB2GRAY}
+                                    };
+};
+
+// ------------ In Range Filter ---------- //
+
+class InRangeFilter : public Filter
+{
+    public:
+        InRangeFilter(string &inRangeName);
+        ~InRangeFilter() {};
+
+        void setParams(XmlRpcValue &params);
+    
+        void apply(Mat &image);
+        void apply(const Mat &src, Mat &dst);
+    
+    private:
+        Scalar upperBounds;
+        Scalar lowerBounds;
+};
+
+// ---------- Median Blur Filter ----------- //
+
+class MedianBlurFilter : public Filter
+{
+    public:
+        MedianBlurFilter();
+        ~MedianBlurFilter() {};
+
+        void setParams(XmlRpcValue &params);
+        void apply(Mat &image);
+        void apply(const Mat &src, Mat &dst);
+
+    private:
+        int kernelSize;
+};
+
+// ------------- Open Filter ------------ //
+
+class OpenFilter : public Filter
+{
+    public:
+        OpenFilter();
+        ~OpenFilter() {};
+
+        void setParams(XmlRpcValue &params);
+        
+        void apply(Mat &image);
+        void apply(const Mat &src, Mat &dst);
+    
+    private:
+        int iterations;
+        Size size;
+};
+
+// ------------ Or Filter ------------- //
+
+class OrFilter : public Filter
+{
+    public:
+        OrFilter();
+        ~OrFilter() {};
+
+        void setParams(XmlRpcValue &params);
+
+        void apply(Mat &image);
+        void apply(const Mat &src, Mat &dst);
+
+    private:
+        vector<Filter *> children;
 };
 
 #endif //FILTER_HPP
